+++ tests/RedisSortedSetsTest.php
@@ -22,185 +22,6 @@ class RedisSortedSetsTest extends TestCase
         $this->keyOptional = 'SortedSets:Optional';
     }
 
-    /*
-     * ========================================================================
-     * zInter
-     * ========================================================================
-     */
-
-    /** @test */
-    public function redis_sorted_sets_zinter_aggregate_max()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.5, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [1, 1], 'MAX'));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-    /** @test */
-    public function redis_sorted_sets_zinter_aggregate_min()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.5, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [1, 1], 'MIN'));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zinter_aggregate_sum()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.0, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [1, 1], 'SUM'));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zinter_multiply_weights()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.0, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [2, 4]));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zinter_with_weights()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.0, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [1, 1]));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zinter_simple_intersection()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.0, 'B'));
-        $this->assertEquals(2, $this->redis->zCard($this->key));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 2.0, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 3.0, 'C'));
-        $this->assertEquals(2, $this->redis->zCard($this->keyOptional));
-
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional]));
-        $this->assertEquals(1, $this->redis->zCard($destinationKey));
-
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-        $this->assertEquals(1, $this->redis->delete($destinationKey));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zinter_wrong_operation()
-    {
-        $destinationKey = $this->key . ':' . $this->keyOptional;
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->keyOptional));
-
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 10.0, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->keyOptional, 5.0, 'B'));
-        // T E S T  -----------------------------------------------------------
-        $this->expectException(SetOperationException::class);
-        $this->redis->zInter($destinationKey, [$this->key, $this->keyOptional], [1, 1], 'AVG');
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->delete($this->keyOptional));
-    }
-
 
     /*
      * ========================================================================
